Remove redundant references to non-existent converters.
It is strange that the project compiles normally. The problem manifested itself only when it was actually launched on a physical device. In debug mode on a physical device, everything worked without errors! I won't be able to get a bug report and had to look for this problem by elimination.